% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CashFlownAnalysis.R
\name{cash_flow}
\alias{cash_flow}
\title{This function generates end of period cash flows,
for receiving returns or paying off debts}
\usage{
cash_flow(prin, i = 0.05, n = 12)
}
\arguments{
\item{prin}{A number, the initial amount or principal}

\item{i}{A decimal, the interest rate per period}

\item{n}{An integer, the number of periods to be studied}
}
\description{
This function generates end of period cash flows,
for receiving returns or paying off debts
}
