% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CashFlownAnalysis.R
\name{cash_viz}
\alias{cash_viz}
\title{This function generates a cash flow visualization}
\usage{
cash_viz(flow)
}
\arguments{
\item{flow}{a vector of the cash flow data}
}
\value{
a cash flow viz
}
\description{
This function generates a cash flow visualization
}
