% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlChart.R
\name{ControlChart}
\alias{ControlChart}
\title{Generate a Control Chart for data}
\usage{
ControlChart(dat, value)
}
\arguments{
\item{dat}{a dataframe}

\item{value}{the column in the dataframe where the values are to graph}
}
\value{
a basic Control Chart visualization
}
\description{
Generate a Control Chart for data
}
