% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFormat.R
\name{ie_viz}
\alias{ie_viz}
\title{This function formats visuals for using general rules}
\usage{
ie_viz(viz, title = NULL, subtitle = NULL, x = NULL, y = NULL)
}
\arguments{
\item{viz}{the visualization object, created with ggplot}

\item{title}{a string for the title of the viz}

\item{subtitle}{a string for the subtitle of the viz}

\item{x}{a string for the title of the x axis of the viz}

\item{y}{a string for the title of the y axis of the viz}
}
\value{
a visualization
}
\description{
This function formats visuals for using general rules
}
