% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRP.R
\name{produce_compare}
\alias{produce_compare}
\title{This function compares different production strategies to meet forcasted demand}
\usage{
produce_compare(forecast, method = "Level", inv0 = 0)
}
\arguments{
\item{forecast}{a vector of forecasted demand for a single product}

\item{method}{a specific production strategy, either Level, Chase, or Flexibility}

\item{inv0}{starting inventory}
}
\value{
a visual showing how inventory and production changes over time
}
\description{
This function compares different production strategies to meet forcasted demand
}
