% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTest.R
\name{stat_anova}
\alias{stat_anova}
\title{Generates an ANOVA table for a given dataset with the associated P-values.}
\usage{
stat_anova(dat, trtA, trtB = NULL, value, twoway = FALSE)
}
\arguments{
\item{dat}{a dataframe, each treatment must be labeled, each row is an observation}

\item{trtA}{the name of the first column that labels each treatment}

\item{trtB}{the name of the first column that labels each treatment, not needed for one way anova}

\item{value}{the name of the column that holds the values}

\item{twoway}{logical, is this a two way ANOVA or one way}
}
\description{
Generates an ANOVA table for a given dataset with the associated P-values.
}
