% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTest.R
\name{stat_ci}
\alias{stat_ci}
\title{Create a Two Sided Confidence Interval for a dataset}
\usage{
stat_ci(dat, per = 0.95)
}
\arguments{
\item{dat}{a list of numerical data}

\item{per}{what percent interval for CI, given as a decimal value}
}
\value{
A vector that gives the upper and lower confidence intervals
}
\description{
Create a Two Sided Confidence Interval for a dataset
}
