% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTest.R
\name{stat_ttest}
\alias{stat_ttest}
\title{Performs a t test on a dataset to see if there is significant difference between means}
\usage{
stat_ttest(dat, trt, value)
}
\arguments{
\item{dat}{a dataframe, each treatment must be labeled, each row is an observation}

\item{trt}{the name of the column that labels each treatment}

\item{value}{the name of the column that holds the values}
}
\description{
Performs a t test on a dataset to see if there is significant difference between means
}
