% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectTimeStudy.R
\name{stnd_time}
\alias{stnd_time}
\title{This Function takes data from a Time Study and finds the Standard Time}
\usage{
stnd_time(dat, irreg = NULL, allow = 0.1)
}
\arguments{
\item{dat}{a dataframe, the observed times for each element of the job, with the accompanying PRs}

\item{irreg}{a dataframe, the irregular job elements, times and frequencies}

\item{allow}{a decimal, the Allowance for this work environment}
}
\description{
This Function takes data from a Time Study and finds the Standard Time
}
