% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WH_Location.R
\name{wh_partition}
\alias{wh_partition}
\title{This function solves a Set Partitioning Model}
\usage{
wh_partition(dat)
}
\arguments{
\item{dat}{A matrix of which Warehouse covers which retailer,
retailers are rows and Warehouses are columns,
1 means the WH can cover the retailer, 0 otherwise}
}
\value{
a matrix of the best result, which WH covers which retailers
}
\description{
This function solves a Set Partitioning Model
}
