% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFormat.R
\name{graphFormat}
\alias{graphFormat}
\title{This function formats visuals for using general rules}
\usage{
graphFormat(viz)
}
\arguments{
\item{viz}{the visualization object, created with ggplot}
}
\value{
a visualization
}
\description{
This function formats visuals for using general rules
}
