% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFormat.R
\name{graphFormat}
\alias{graphFormat}
\title{This function formats visuals for using general rules}
\usage{
graphFormat()
}
\value{
a visualization
}
\description{
This function formats visuals for using general rules
}
